Updated the music playback functions.
There were a few problems with the way it was defined previously. To start with, music wouldn't play on Linux because the paths were defined in a Windows-standard way. So I updated the playback to use os.path.join to define the location of the music files. Next I discovered that all of the music tracks would automatically start playing when the game was started, and when stopped, couldn't start playing again, because rather than being started when the menu option was clicked, they were being started when the menu options were created. So I fixed that by using functools.partial. Next, I cleaned up the play_music_track function to play any specified track, rather than having one function per track, and I set up the menu so that it can load any number of tracks rather than having to define each one individually in the menu construction. All of the tracks are now defined by their mp3 filename in the `available_songs` dictionary at the top of the source file. To add new songs to the application, simply add the mp3 to the music folder and add an entry to the `available_songs` variable.