Consolidated functions that open a web browser.
One of the cardinal rules is Don't Repeat Yourself (DRY). If you have multiple functions that do the same basic thing, make them into one function that can handle all your processes. I combined both of the functions that open browsers into a single function, so that in the future, if we need to open another browser for any reason, we can just re-use that function again instead of writing a new one. (The function is called `open_browser` and takes the URL string as the only argument.)